refactor(routes): split routes.ts into per-surface modules (#311)
routes.ts was a 2631-line buildRouter closure spanning auth, invites,
admin, groups, sessions/ports/tabs/uploads, and templates. Split each
surface into routes/<surface>.ts exporting `registerX(router, ctx)`, with
a shared routes/shared.ts holding the RouteContext + cross-surface helpers
(serializeMeta, serializeUsage, handleSessionError, TERMINAL_DIM_MAX).

buildRouter is now a thin orchestrator: it builds one RouteContext and
calls the register functions in the original registration order — which is
load-bearing (the auth-prefix `requireAuth` mounts and the idle-bump
middleware must sit between the public auth routes and the authed
surfaces, and Express matches routes in registration order).

Behaviour-preserving: handler bodies were moved verbatim (each register
fn re-binds the limiters/deps from ctx to the same local names the bodies
already used). Public API is unchanged — buildRouter, TERMINAL_DIM_MAX,
TemplateBodyError, assertTemplateConfigShape, parseTemplateBody are still
exported from routes.ts (the latter via re-export).

routes.ts: 2631 -> 132 lines. Full suite (905) green, tsc + Biome clean.
